moved quiz files to components folder to clean up file structure
