+++ src/components/pages/French.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import Quiz from 'react-quiz-component';
-import quiz from './Frenchquiz';
+import quiz from '../Frenchquiz.json';
 import PageTitle from "../PageTitle";
 import { FlashcardArray } from "react-quizlet-flashcard";
 import frenchData from '../French.json'
@@ -17,12 +17,13 @@ import Blanc from '../../sounds/Frenchsounds/Blanc.mp3';
 import Brun from '../../sounds/Frenchsounds/Brun.mp3';
 
 function French() {
-
+  //setting state values
   const [showFlash, setShowFlash] = useState(true);
   const [currentCard, setCurrentCard] = useState(1);
   const [data] = useState([...frenchData]);
   const [categoryFilter, setCategoryFilter] = useState('All');
 
+  //function to filter json data by category
   const filteredData = data.filter(item => {
     if (categoryFilter === 'All') {
       return true;
@@ -31,18 +32,22 @@ function French() {
     }
   });
 
+  //function to determine when the flashcards show up
   const handleFlashClick = () => {
     setShowFlash(true);
   };
 
+  //function to determine when the quiz shows up
   const handleQuizClick = () => {
     setShowFlash(false);
   };
 
+  //function to determine which category was selected
   const handleCategoryClick = (category) => {
     setCategoryFilter(category);
   };
 
+  //function to play audio clips depending on current card index
   function play() {
     if (currentCard === 1) {
     new Audio(Rouge).play()}
@@ -66,6 +71,7 @@ function French() {
       new Audio(Gris).play()
     }
 
+    //function to update card index
     const handleCardChange = (index) => {
       setCurrentCard(index);
     };
